% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{hansard_fragment}
\alias{hansard_fragment}
\title{Obtain a HTML fragment of a Hansard record}
\usage{
hansard_fragment(pdfid)
}
\arguments{
\item{pdfid}{Document ID}
}
\value{
full response (invisibly)
}
\description{
Obtain a HTML fragment of a Hansard record
}
\examples{
\dontrun{
hansard_fragment(pdfid = 'HANSARD-10-19961')
}
}

