% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{hansard_toc}
\alias{hansard_toc}
\title{Obtain a Hansard Table of Contents}
\usage{
hansard_toc(docid)
}
\arguments{
\item{docid}{Document ID}
}
\value{
full response (invisibly)
}
\description{
Obtain a Hansard Table of Contents
}
\examples{
hansard_toc(docid = 'HANSARD-10-19980')
}

