% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{hansard_year}
\alias{hansard_year}
\title{Obtain the year's Hansard records}
\usage{
hansard_year(year)
}
\arguments{
\item{year}{year for which the records should be obtained}
}
\value{
full response (invisibly)
}
\description{
Obtain the year's Hansard records
}
\examples{
hansard_year(year = "2017")
}

